backend/hikeit-main/requirements.txt
asgiref==3.6.0
astroid==2.15.0
black==23.1.0
# btrfsutil==6.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cloudinary==1.32.0
cryptography==39.0.2
dill==0.3.6
Django==4.1.6
django-extensions==3.2.1
django-phonenumber-field==7.0.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
geographiclib==2.0
geopy==2.3.0
gunicorn==20.1.0
idna==3.4
isort==5.12.0
jedi==0.18.2
lazy-object-proxy==1.9.0
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
parso==0.8.3
pathspec==0.11.0
Pillow==9.4.0
platformdirs==3.0.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycairo==1.23.0
pycparser==2.21
PyGObject==3.42.2
PyJWT==2.6.0
pylint==2.17.0
pyOpenSSL==23.0.0
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
TBB==0.2
team==1.0
tomli==2.0.1
tomlkit==0.11.6
twilio==7.16.4
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
wrapt==1.15.0
